+++ sdk/core/core-rest-pipeline/test/bearerTokenAuthenticationPolicy.spec.ts
@@ -12,9 +12,8 @@ import {
   bearerTokenAuthenticationPolicy,
   SendRequest
 } from "../src";
-import { DEFAULT_CYCLER_OPTIONS } from "../src/util/tokenCycler";
 
-const { refreshWindowInMs: defaultRefreshWindow } = DEFAULT_CYCLER_OPTIONS;
+const defaultRefreshWindow = 1000 * 60 * 2; // Start refreshing 2m before expiry
 
 describe("BearerTokenAuthenticationPolicy", function() {
   let clock: sinon.SinonFakeTimers;
@@ -58,100 +57,12 @@ describe("BearerTokenAuthenticationPolicy", function() {
     assert.strictEqual(request.headers.get("Authorization"), `Bearer ${mockToken}`);
   });
 
-  it("refreshes the token on initial request", async () => {
-    const expiresOn = Date.now() + 1000 * 60; // One minute later.
-    const credential = new MockRefreshAzureCredential(expiresOn);
-    const request = createPipelineRequest({ url: "https://example.com" });
-
-    const successResponse: PipelineResponse = {
-      headers: createHttpHeaders(),
-      request,
-      status: 200
-    };
-    const next = sinon.stub<Parameters<SendRequest>, ReturnType<SendRequest>>();
-    next.resolves(successResponse);
-
-    const policy = createBearerTokenPolicy("test-scope", credential);
-
-    await policy.sendRequest(request, next);
-    assert.strictEqual(credential.authCount, 1);
-  });
-
-  it("refreshes the token during the refresh window", async () => {
-    const expireDelayMs = defaultRefreshWindow + 5000;
-    let tokenExpiration = Date.now() + expireDelayMs;
-    const credential = new MockRefreshAzureCredential(tokenExpiration);
-
-    const request = createPipelineRequest({ url: "https://example.com" });
-    const successResponse: PipelineResponse = {
-      headers: createHttpHeaders(),
-      request,
-      status: 200
-    };
-    const next = sinon.stub<Parameters<SendRequest>, ReturnType<SendRequest>>();
-    next.resolves(successResponse);
-
-    const policy = createBearerTokenPolicy("test-scope", credential);
-
-    // The token is cached and remains cached for a bit.
-    await policy.sendRequest(request, next);
-    await policy.sendRequest(request, next);
-    assert.strictEqual(credential.authCount, 1);
-
-    // The token will remain cached until tokenExpiration - testTokenRefreshBufferMs, so in (5000 - 1000) milliseconds.
-
-    // For safe measure, we test the token is still cached a second earlier than the forced refresh expectation.
-    clock.tick(expireDelayMs - defaultRefreshWindow - 1000);
-    await policy.sendRequest(request, next);
-    assert.strictEqual(credential.authCount, 1);
-
-    // The new token will last for a few minutes again.
-    tokenExpiration = Date.now() + expireDelayMs;
-    credential.expiresOnTimestamp = tokenExpiration;
-
-    // Now we wait until it expires:
-    clock.tick(expireDelayMs + 1000);
-    await policy.sendRequest(request, next);
-    assert.strictEqual(credential.authCount, 2);
-  });
-
-  it("access token refresher should prevent multiple initial getToken requests to break", async () => {
-    const expireDelayMs = 5000;
-    const startTime = Date.now();
-    const tokenExpiration = startTime + expireDelayMs;
-    const getTokenDelay = 100;
-    const credential = new MockRefreshAzureCredential(tokenExpiration, getTokenDelay, clock);
-
-    const request = createPipelineRequest({ url: "https://example.com" });
-    const successResponse: PipelineResponse = {
-      headers: createHttpHeaders(),
-      request,
-      status: 200
-    };
-    const next = sinon.stub<Parameters<SendRequest>, ReturnType<SendRequest>>();
-    next.resolves(successResponse);
-
-    const policy = createBearerTokenPolicy("test-scope", credential);
-
-    // Now we send some requests.
-    const promises = [
-      policy.sendRequest(request, next),
-      policy.sendRequest(request, next),
-      policy.sendRequest(request, next)
-    ];
-    // Now we wait until they're all resolved.
-    for (const promise of promises) {
-      await promise;
-    }
-    assert.strictEqual(credential.authCount, 1, "The first authentication should have happened");
-  });
-
   it("credential errors should bubble up", async () => {
     const expireDelayMs = 5000;
     const startTime = Date.now();
     const tokenExpiration = startTime + expireDelayMs;
     const getTokenDelay = 100;
-    const credential = new MockRefreshAzureCredential(tokenExpiration, getTokenDelay, clock);
+    const credential = new MockAzureCredential(tokenExpiration, getTokenDelay, clock);
 
     const request = createPipelineRequest({ url: "https://example.com" });
     const successResponse: PipelineResponse = {
@@ -181,52 +92,10 @@ describe("BearerTokenAuthenticationPolicy", function() {
     );
   });
 
-  it("access token refresher should prevent refreshers to happen too fast while the token is about to expire", async () => {
-    const expireDelayMs = 5000;
-    const startTime = Date.now();
-    const tokenExpiration = startTime + defaultRefreshWindow + expireDelayMs;
-    const getTokenDelay = 100;
-    const credential = new MockRefreshAzureCredential(tokenExpiration, getTokenDelay, clock);
-
-    const request = createPipelineRequest({ url: "https://example.com" });
-    const successResponse: PipelineResponse = {
-      headers: createHttpHeaders(),
-      request,
-      status: 200
-    };
-    const next = sinon.stub<Parameters<SendRequest>, ReturnType<SendRequest>>();
-    next.resolves(successResponse);
-
-    const policy = createBearerTokenPolicy("test-scope", credential);
-
-    await policy.sendRequest(request, next);
-    assert.strictEqual(credential.authCount, 1, "The first authentication should have happened");
-
-    clock.tick(tokenExpiration - startTime - defaultRefreshWindow); // Until we start refreshing the token
-
-    // Now we wait until some requests are all resolved.
-    await Promise.all([
-      policy.sendRequest(request, next),
-      policy.sendRequest(request, next),
-      policy.sendRequest(request, next)
-    ]);
-
-    // Only getTokenDelay should have passed, and only one refresh should have happened.
-    assert.strictEqual(
-      credential.authCount,
-      2,
-      "authCode should have been called once during the refresh time"
-    );
-
-    const exceptionMessage =
-      "the total time passed should be in the refresh room, plus the many getTokens that have happened so far";
-    assert.equal(expireDelayMs + 2 * getTokenDelay, Date.now() - startTime, exceptionMessage);
-  });
-
   it("throws if the target URI doesn't start with 'https'", async () => {
     const expireDelayMs = defaultRefreshWindow + 5000;
     const tokenExpiration = Date.now() + expireDelayMs;
-    const credential = new MockRefreshAzureCredential(tokenExpiration);
+    const credential = new MockAzureCredential(tokenExpiration);
 
     const request = createPipelineRequest({ url: "http://example.com" });
     const policy = createBearerTokenPolicy("test-scope", credential);
@@ -256,7 +125,7 @@ describe("BearerTokenAuthenticationPolicy", function() {
   }
 });
 
-class MockRefreshAzureCredential implements TokenCredential {
+class MockAzureCredential implements TokenCredential {
   public authCount = 0;
   public shouldThrow: boolean = false;